resharped... now it's cleaner and meaner.
BUT, the only TCP messages coming out of the ROSCPP publisher are keepalives... maybe I need to be a conformist and publish rosout like all of the linux nodes?

+++ ROS#/EricIsAMAZING/Header.cs
@@ -24,7 +24,7 @@ namespace EricIsAMAZING
                 if (chunks.Length != 2)
                     return false;
                 Values[chunks[0].Trim()] = chunks[1].Trim();
-                i += (int) size;
+                i += size;
             }
             return true;
         }
